Move e2e tests into self-contained systems and delete End2EndTests project
