--- application/End2EndTests/End2EndTests.esproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.VisualStudio.JavaScript.Sdk/1.0.751930">
-
-    <PropertyGroup>
-        <ShouldRunNpmInstall>false</ShouldRunNpmInstall>
-        <ShouldRunBuildScript>false</ShouldRunBuildScript>
-        <TargetFrameworkMoniker>net9.0</TargetFrameworkMoniker>
-        <DefaultItemExcludes>$(DefaultItemExcludes);node_modules\**;dist\**;playwright-report\**;test-results\**;*.config.*;*.d.ts</DefaultItemExcludes>
-        <NpmInstallCheck>$(MSBuildProjectDirectory)\..\package-lock.json</NpmInstallCheck>
-    </PropertyGroup>
-
-</Project>
