+++ application/PlatformPlatform.slnx
@@ -20,5 +20,4 @@
     </Folder>
     <Project Path="AppGateway\AppGateway.csproj" />
     <Project Path="AppHost\AppHost.csproj" />
-    <Project Path="End2EndTests\End2EndTests.esproj" />
 </Solution>
