+++ application/account-management/WebApp/tests/e2e/login.spec.ts
@@ -5,8 +5,8 @@ import {
   assertToastMessage,
   assertValidationError,
   createTestContext
-} from "../../../shared-webapp/tests/e2e/utils/test-assertions";
-import { getVerificationCode, testUser } from "../../../shared-webapp/tests/e2e/utils/test-data";
+} from "../../../../shared-webapp/tests/e2e/utils/test-assertions";
+import { getVerificationCode, testUser } from "../../../../shared-webapp/tests/e2e/utils/test-data";
 
 test.describe("Login", () => {
   test.describe("@smoke", () => {
